Move dummyUser to DummyData folder
