+++ src/DummyData/DummyUser.js
@@ -1,0 +1,33 @@
+const dummyUser = {
+	username: "username",
+	fullname: "Name Surname",
+	twitter: "https://github.com/adnoto-team",
+	totalComments: 0,
+	desc:
+		"Lorem ipsum, dolor sit amet consectetur adipisicing elit. Repellendus, quas quisquam sapiente maxime at in magnam accusantium officiis fugit ad cumque nemo vero fugiat soluta est nobis? Eaque, rem pariatur!",
+	avatarImg: "https://supercharge.info/images/avatar-placeholder.png",
+	backgroundImage:
+		"https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fjessicajjohnston.com%2Fblog%2Fwp-content%2Fuploads%2F2014%2F04%2Fthe-office-netflix-tv-series.jpg&f=1&nofb=1",
+	currentlyWatching: [
+		{
+			title: "The Office",
+			imdb: "8.5",
+			coverImg:
+				"https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Floftcinema.org%2Ffiles%2F2017%2F08%2F740full-the-office-us-poster.jpg&f=1&nofb=1",
+		},
+		{
+			title: "Seinfeld",
+			imdb: "10",
+			coverImg:
+				"https://m.media-amazon.com/images/M/MV5BZjZjMzQ2ZmUtZWEyZC00NWJiLWFjM2UtMzhmYzZmZDcxMzllXkEyXkFqcGdeQXVyNTA4NzY1MzY@._V1_UY268_CR3,0,182,268_AL_.jpg",
+		},
+		{
+			title: "Friends",
+			imdb: "9.5",
+			coverImg:
+				"https://m.media-amazon.com/images/M/MV5BNDVkYjU0MzctMWRmZi00NTkxLTgwZWEtOWVhYjZlYjllYmU4XkEyXkFqcGdeQXVyNTA4NzY1MzY@._V1_UY268_CR0,0,182,268_AL_.jpg",
+		},
+	],
+};
+
+export default dummyUser;
