+++ src/DummyData/movies.js
@@ -1,49 +1,49 @@
 const Data = [
-    {
-        name: "The Movie (2020)",
-        imdbScore: 4,
-        thumbnail: "https://d13ezvd6yrslxm.cloudfront.net/wp/wp-content/images/potterebellposter.jpg"
-    },
-    {
-        name: "The Movie (2020)",
-        imdbScore: 4,
-        thumbnail: "https://www.joblo.com/assets/images/joblo/posters/2019/01/IO-poster-1.jpg"
-    },
-    {
-        name: "The Movie (2020)",
-        imdbScore: 4,
-        thumbnail: "https://i.imgur.com/RoF2unX.jpg"
-    },
-    {
-        name: "The Movie (2020)",
-        imdbScore: 4,
-        thumbnail: "https://cdn11.bigcommerce.com/s-ydriczk/images/stencil/1280x1280/products/89058/93685/Joker-2019-Final-Style-steps-Poster-buy-original-movie-posters-at-starstills__62518.1572351179.jpg?c=2?imbypass=on"
-    },
-    {
-        name: "The Movie (2020)",
-        imdbScore: 4,
-        thumbnail: "https://d13ezvd6yrslxm.cloudfront.net/wp/wp-content/images/potterebellposter.jpg"
-    },
-    {
-        name: "The Movie (2020)",
-        imdbScore: 4,
-        thumbnail: "https://www.joblo.com/assets/images/joblo/posters/2019/01/IO-poster-1.jpg"
-    },
-    {
-        name: "The Movie (2020)",
-        imdbScore: 4,
-        thumbnail: "https://i.imgur.com/RoF2unX.jpg"
-    },
-    {
-        name: "The Movie (2020)",
-        imdbScore: 4,
-        thumbnail: "https://cdn11.bigcommerce.com/s-ydriczk/images/stencil/1280x1280/products/89058/93685/Joker-2019-Final-Style-steps-Poster-buy-original-movie-posters-at-starstills__62518.1572351179.jpg?c=2?imbypass=on"
-    },
- 
- 
-    
-  
-      
-    
-]
-export default Data
+	{
+		name: "The Movie (2020)",
+		imdbScore: 4,
+		thumbnail:
+			"https://d13ezvd6yrslxm.cloudfront.net/wp/wp-content/images/potterebellposter.jpg",
+	},
+	{
+		name: "The Movie (2020)",
+		imdbScore: 4,
+		thumbnail:
+			"https://www.joblo.com/assets/images/joblo/posters/2019/01/IO-poster-1.jpg",
+	},
+	{
+		name: "The Movie (2020)",
+		imdbScore: 4,
+		thumbnail: "https://i.imgur.com/RoF2unX.jpg",
+	},
+	{
+		name: "The Movie (2020)",
+		imdbScore: 4,
+		thumbnail:
+			"https://cdn11.bigcommerce.com/s-ydriczk/images/stencil/1280x1280/products/89058/93685/Joker-2019-Final-Style-steps-Poster-buy-original-movie-posters-at-starstills__62518.1572351179.jpg?c=2?imbypass=on",
+	},
+	{
+		name: "The Movie (2020)",
+		imdbScore: 4,
+		thumbnail:
+			"https://d13ezvd6yrslxm.cloudfront.net/wp/wp-content/images/potterebellposter.jpg",
+	},
+	{
+		name: "The Movie (2020)",
+		imdbScore: 4,
+		thumbnail:
+			"https://www.joblo.com/assets/images/joblo/posters/2019/01/IO-poster-1.jpg",
+	},
+	{
+		name: "The Movie (2020)",
+		imdbScore: 4,
+		thumbnail: "https://i.imgur.com/RoF2unX.jpg",
+	},
+	{
+		name: "The Movie (2020)",
+		imdbScore: 4,
+		thumbnail:
+			"https://cdn11.bigcommerce.com/s-ydriczk/images/stencil/1280x1280/products/89058/93685/Joker-2019-Final-Style-steps-Poster-buy-original-movie-posters-at-starstills__62518.1572351179.jpg?c=2?imbypass=on",
+	},
+];
+export default Data;
